changed float types to long

+++ src/ADALottoGameWinner.cs
@@ -7,7 +7,7 @@ namespace ADALottoModels
     public class ADALottoGameWinner
     {
         public string Address { get; set; } = string.Empty;
-        public float DrawBlockId { get; set; }
+        public long DrawBlockId { get; set; }
         public DateTime DrawDate { get; set; }
         public double Prize { get; set; }
         public Transaction? RewardTx { get; set; }
